LayoutTests/http/tests/misc/webtiming-resolution-expected.txt

Verifies the minimum resolution is 5 microseconds.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ((t1 - t0) / 0.005) % 1 < 1e-10 is true
